Backend/requirements.txt
Django
djangorestframework
django-cors-headers
docx
python-docx
djangorestframework
reportlab